+++ lib/time_machine/validation/validator_factory.rb
@@ -13,6 +13,7 @@ require './lib/time_machine/validators/geom_invalid'
 require './lib/time_machine/validators/geom_new_object'
 require './lib/time_machine/validators/network'
 require './lib/time_machine/validators/tags_changes'
+require './lib/time_machine/validators/user_block'
 require './lib/time_machine/validators/user_list'
 require './lib/time_machine/validators/validator_link'